refactor: refactor name files and dir, move type
