--- src/components/game/panel-control.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const PanelGameControl = () => {
-  return <div>Panel Control</div>;
-};
-
-export default PanelGameControl;
