--- src/components/game/panel-dealer.tsx
@@ -1,8 +1,0 @@
-import React from "react";
-
-const PanelDealer = () => {
-  return <div>Panel Dealer</div>;
-};
-
-export default PanelDealer;
-
